Create FilterLink
We will start by creating a new FilterLink component that the user will
click to switch the current visible items. This component accepts the
filter prop (just a string) & children (the contents of the link).

The component will be a simple <a> tag that when clicked will dispatch
an action of type 'SET_VISIBILITY_FILTER' along with a filter prop so
the reducer knows which filter is being clicked.

We pass the children down to the <a> tag so the text of the link can be
specified.