Change active filter link to span
Add currentFilter as a prop to FilterLink.

Add a condition that says when the filter is the current filter, it
will become static text.

—

Clicking one of the filter types dispatches an action of type
'SET_VISIBILITY_FILTER' and passes filter which is a prop to the
FilterLink component (every one of the 3 links will have a different
filter prop).

The store.dispatch() function calls our todoApp() root reducer with the
state and the action, which in turn calls the visibilityFilter()
reducer with the part of the state and the action.

Note that when the action is of type 'SET_VISIBILITY_FILTER', it
doesn't care about the previous state. It just returns action.filter as
the next state of the visibilityFilter() reducer. The root reducer will
use this new field as part of its new state object.

Because the render() function is subscribed to changes in store, it's
going to get the new state object and pass all its keys as props to the
TodoApp component.

The TodoApp component receives all the todos as well as the newly
updated visibilityFilter as its props, which are then passed to the
getVisibleTodos() function The currently visible todos are calculated
based on the current visibilty filter ('SHOW_ALL', 'SHOW_COMPLETED', or
'SHOW_ACTIVE').

Depending on which filter is selected, getVisibleTodos() may return a
brand new array of todos containing only the appropriate items. This
returned array is then enumerated and rendered inside of TodoApp's
render() function.

The visibilityFilter field is also used by the FilterLinks as the
currentFilter because the FilterLink wants to know if its filter is the
current one so it can render the style appropriately (i.e. the current
filter is active, so the user shouldn't be able to click it).