Extract Todo into presentational component
It's best practice with React to have several components that don't
specify any behaviors, and only are concerned with how things are
rendered (how they look).

Because we want our list to be a presentational component, we "promote"
the onClick handler to become a prop.

We also want to be more explicit about what the data is that the
component needs to render. Instead of passing a todo object, we will
pass completed and text fields as separate props.