Create Footer functional component
Presentational components

Since each FilterLink needs to know the visibilityFilter, we will make
that a prop.

Replace FilterLink’s store.dispatch with an onClick call that will take
a single parameter with the filter.

Since onClick is now a prop for FilterLink, we need to specify it every
time that FilterLink is used in our Footer. Adding
onClick={onFilterClick} makes sure that onClick makes it to FilterLink
as a prop.