Change TodoApp into a function
This allows us to eliminate the destructuring of todos and
visibilityFilter from this.props inside the render function. Instead,
we can do this inside the argument to the TodoApp function.

We can also do away with the render() declaration.

Since the visibleTodos are only used in a single place, we can move its
declaration into the TodoList todos prop declaration.

One of the downsides to having separate presentational components is
that we have to move around a lot of props, including the callbacks.