Previously, the FilterLink needed to know the current filter so it could change its appearance when it's active. However, in order for it to receive the current filter, it had to be passed down from the top. This is why Footer had to be given visibilityFilter so it can be passed to a FilterLink.
Previously the Footer component accepted the visibilityFilter and
onFilterClick() callback as its props, but it doesn't actually use
either of them. It just passes down to the FilterLink. We can only do
this because we know that the Footer component doesn't care about the
values of its props, as they only exist to pass down to FilterLink.

We start by removing the props definition from the Footer component,
and removing them from the FilterLinks as well.