Create FilterLink container component
The new FilterLink will be a class that renders the Link with the
current data from the store. It's going to read the component props and
read the state.

As a container component, FilterLink doesn't have its own markup, and
it delegates rendering to the Link presentational component. In this
case, it calculates its active prop by comparing its own filter prop
with the visibilityFilter in the Redux store's state.

The filter prop is the one that is passed to the FilterLink from the
Footer. The visibilityFilter corresponds to the current chosen
visibility filter that is held in Redux store's state. If they match,
we want the link to appear active.

The container component also needs to specify the behavior. In this
case, the FilterLink specifies that when a particular Link is clicked,
we should dispatch an action of the type 'SET_VISIBILITY_FILTER' along
with the filter value that we take from the props.

The FilterLink may accept children, which will be used as the contents
of the Link. So we are going to pass the children down to the Link
component, which is going to render them inside of the <a> tag.