Force re-render FilterLink when store changes
Previously FilterLink read the state of the Redux store but did not
subscribe to it. If the parent component (Link) doesn’t update the
store when the store is updated, the correct value won’t be rendered.

React provides a special forceUpdate() method on the Component
instances to force them to re-render. We can use it in combination with
the store.subscribe() method so that any time the store changes we
force the container component to update.