Create VisibleTodoList container component
So that the TodoList component can remain purely presentational, we are
encapsulate reading the currently visible todos into a container
component (VisableTodoList).

Data for the VisibleTodoList will be calculated by using the current
state.

We use getVisibleTodos() to determine which todos to show based on the
visibilityFilter.

The VisibleTodoList container component also specifies the behavior of
onTodoClick() and dispatch the ‘TOGGLE_TODO’ action.

We force re-render the VisibleTodoList just like the FilterLink
component because If the parent component (VisibleTodoList) doesn’t
update the
store when the store is updated, the correct todos won’t be rendered.
The component must subscribe to the store.

*Container components such as FilterLink and connect a presentational
component to the Redux store, and specify the data and behavior that it
needs.