Refactor TodoApp to use the container components
None of the containers need props from TodoApp (I think it is because
they each subscribe to the store).

The container components take care of re-rendering when the store
changes.

Data Flow:
* Only 1 ReactDOM.render call- we don’t have to re-render again when
the store changes because the container components take care of that.
* AddTodo: not considered a purely presentational nor a container
component. The input and button are presentational but dispatching the
action is the behavior (usually specified by the container). We keep to
together because there isn’t any state, the UI is very simple, and it’s
unlikely to have any behavior other than dispatching the add todo
action.
* VisibleTodoList: container component that subscribe to the store and
re-renders the todolist anytime the store state changes. Calculates the
visible todos from the redux store state and passes them as the todos.
When the todos are clicked, it dispatches the TOGGLE_TODO action.
Rendering is done by the TodoList component that renders the todos
passed to it and binds their clicks to the onTodoClick prop.
* Footer: Presentational component that renders 3 filter links.
* FilterLink(within footer): container component that subscribes to the
store and render the presentational component (Link). It determines if
Link is active based on its props and the redux store state  and
specifics the behavior (what happens when the Link is clicked).
* Link: presentational component that renders an anchor tag.

Recommendation:
1. Extract the presentational components.
2. If there is too much boilerplate passing the props to them, create
the container components around them that load the data and specify the
behavior.