Inject store into TodoApp
Previously, once the store had been created, the container components
got data from it by calling store.getState(), and subscribed to changes
by calling store.subscribe(), and dispatch actions by calling
store.dispatch().

Having all of our code in a single file works well for a simple
example, but it doesn't scale very well:
1. It makes our components harder to test because they'll reference a
specific store, but we may want to provide a different mock store in
the tests.
2. It makes it hard to implement Universal applications that are
rendered on the server, because on the server we'll want to supply a
different store instance for every request, because different requests
have different data.