Refactor TodoApp to accept store
Every container component needs a reference to store, and unfortunately
the only way to make this happen (for now!) is by passing it down to
every component as a prop.

The problem with doing it this way is that the container components
need to have the store instance to get state from it, dispatch actions,
and subscribe to changes.