Adjust container components to to take the store from the props in containerDidMount() and render()
Now all of our components are receiving state via their props instead
of relying on a top level variable.

This change does not change the behavior of the data flow of the
application. The containers subscribe to store and update like they did
before. What changed is how they get the store.

* Note: Though Footer itself does not need the store, we must give it
to Footer so we can pass it to FilterLink