Update Provider component to use React's context feature
This is what makes the store available to any component that it
renders. Now the store will be available to any of the children and
grandchildren of the components Provider renders.

e have to provide childContextTypes on the component that provides
child context. This is similar to React's PropTypes definition that
helps you when writing your app, except that in this case it is
required in order for the child components to receive context.