Pass the Store Down Implicitly via Context
1. Render TodoApp within Provider component.
2. Provider renders whatever you pass to it (TodoApp) and provides
context to any components inside (including grandchildren).
2. Context contains the store which corresponds to the store we pass as
a prop to the provider component.
3. Store is passed to Provider in render call.
4. Store is made available to child by defining getChildContext() with
the store key pointing to the store prop.
5. Provider requires childContextTypes where we specify that the store
key has propType of object, if we want to pass it down. Now we don’t
have to pass the store through the intermediate components. 6. We
declare the context types on the components that need access to the
store so the can retrieve it from the context instead of the props.
7. In AddTodo, we use the deconstruction syntax to grab the store from
the context.
8. Remove context types from footer.
9. Since the FilterLink directly uses the context, it has to define
contextTypes.