connect()
Together, these mapStateToProps & mapDispatchToProps describe a
container component so well that instead of writing it we can generate
it by using the connect() function provided by react-redux.

* Now instead of creating a VisibleTodoList class, we can declare a
variable and use the connect() method to obtain it. We'll pass in
mapStateToProps as the first argument, and mapDispatchToProps as the
second. As this is a curried function, we must call it again, passing
in the presentational component that we want it to wrap and pass the
props, thereby connecting to the redux store (in this case, TodoList).
* The result of the connect call is the container component that is
going to render the presentational component. It will calculate the
props to pass to the presentational component by merging the objects
returned from mapStateToProps, mapDispatchToProps, and its own props.
* The connect() function will generate the component just like the one
we wrote by hand, so we don't have to write the code to subscribe to
the store or to specify the context types manually.