Generate FooterLink
* mapStateToProps accepts the state of the Redux store, and returns the
props that should be passed to the Link component.
* It's common to use the container props when calculating the child
props, so we pass them in as a second argument to mapStateToProps. In
this case, we'll rename it to ownProps to make it more clear that we
are talking about the container component's own props, and not the
props that are passed to the child, which is the return value of
mapStateToProps.
* mapDispatchToProps: In this case, we only have the onClick() where we
dispatch the action of type 'SET_VISIBILITY_FILTER' along with the
filter type. Since there is another reference to props, we will add
ownProps as our second argument to mapDispatchToLinkProps.